requirements.txt
alembic==0.8.8
awesome-slugify==1.6.5
cffi==1.8.2
click==6.6
cryptography==1.5
ecdsa==0.13
enum34==1.1.6
Flask==0.11.1
Flask-Assets==0.12
Flask-Login==0.3.2
Flask-Migrate==2.0.0
Flask-OAuth==0.12
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
httplib2==0.9.2
idna==2.1
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
oauth2==1.9.0.post1
paramiko==2.0.2
pathlib==1.0.1
pkg-resources==0.0.0
psycopg2==2.6.2
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pyScss==1.3.5
python-editor==1.0.1
regex==2016.8.27
six==1.10.0
SQLAlchemy==1.0.15
Unidecode==0.4.19
webassets==0.12.0
Werkzeug==0.11.11
